Include image_enrichment module and supporting code
The new image_enrichment module includes an enrichment class definition
for applying labels to attached Tweet images based on [an open source
implementation](https://keras.io/applications/) of the VGG16 network
architecture. Some of the methods of this class could be abstracted
into a generic base class if/when the need arises, but for now are
not.

The setup is a bit more involved for this enrichment because of the
list of library dependencies, and the one-time downloads of the network
weights and class labels. The included bash script (and requirements
file) is an attempt to make this relatively straightforward for the
user.